+++ src/modules/AnnouncesSpot/data.ts
@@ -20,32 +20,32 @@ const announcesVideos = [
 const announcesTitles: IDataTranslate = {
   'ru': [
     {
-    bigTitle: 'Смотреть на телевизоре.',
-    description: 'Смотрите на Smart TV, PlayStation, Xbox, Chromecast, Apple TV, плеерах Blu-ray и других устройствах.',
-    imgSrc: 'https://assets.nflxext.com/ffe/siteui/acquisition/ourStory/fuji/desktop/tv.png',
-    altDescr: 'just TV',
-  },
-  {
-    bigTitle: 'Загружайте сериалы для просмотра офлайн.',
-    description: 'Сохраняйте видео в избранном, и у вас всегда будет, что посмотреть.',
-    imgSrc: 'https://assets.nflxext.com/ffe/siteui/acquisition/ourStory/fuji/desktop/mobile-0819.jpg',
-    altDescr: 'see movies offline',
-  },
-  {
-    bigTitle: 'Смотрите где угодно.',
-    description:
-      'Смотрите фильмы и сериалы на телефоне, планшете, ноутбуке и телевизоре без ограничений и без дополнительной платы.',
-    imgSrc: 'https://assets.nflxext.com/ffe/siteui/acquisition/ourStory/fuji/desktop/device-pile.png',
-    altDescr: 'any gadget',
-  },
-  {
-    bigTitle: 'Создавайте профили для детей.',
-    description:
-      'Подарите детям мир приключений с их любимыми героями. Он создан специально для них и уже доступен с вашей подпиской.',
-    imgSrc:
-      'https://occ-0-3666-2773.1.nflxso.net/dnm/api/v6/19OhWN2dO19C9txTON9tvTFtefw/AAAABdFTpLmANuJpYneLq8L5m7CunMCi8e8Nl4y7xaPVWzG3IeoDoq17egTQAthApKg_4sdRWdwuR8KadWu1frjL3JQImpwq.png?r=fcd',
-    altDescr: 'for children',
-  },
+      bigTitle: 'Смотреть на телевизоре.',
+      description: 'Смотрите на Smart TV, PlayStation, Xbox, Chromecast, Apple TV, плеерах Blu-ray и других устройствах.',
+      imgSrc: 'https://assets.nflxext.com/ffe/siteui/acquisition/ourStory/fuji/desktop/tv.png',
+      altDescr: 'just TV',
+    },
+    {
+      bigTitle: 'Загружайте сериалы для просмотра офлайн.',
+      description: 'Сохраняйте видео в избранном, и у вас всегда будет, что посмотреть.',
+      imgSrc: 'https://assets.nflxext.com/ffe/siteui/acquisition/ourStory/fuji/desktop/mobile-0819.jpg',
+      altDescr: 'see movies offline',
+    },
+    {
+      bigTitle: 'Смотрите где угодно.',
+      description:
+        'Смотрите фильмы и сериалы на телефоне, планшете, ноутбуке и телевизоре без ограничений и без дополнительной платы.',
+      imgSrc: 'https://assets.nflxext.com/ffe/siteui/acquisition/ourStory/fuji/desktop/device-pile.png',
+      altDescr: 'any gadget',
+    },
+    {
+      bigTitle: 'Создавайте профили для детей.',
+      description:
+        'Подарите детям мир приключений с их любимыми героями. Он создан специально для них и уже доступен с вашей подпиской.',
+      imgSrc:
+        'https://occ-0-3666-2773.1.nflxso.net/dnm/api/v6/19OhWN2dO19C9txTON9tvTFtefw/AAAABdFTpLmANuJpYneLq8L5m7CunMCi8e8Nl4y7xaPVWzG3IeoDoq17egTQAthApKg_4sdRWdwuR8KadWu1frjL3JQImpwq.png?r=fcd',
+      altDescr: 'for children',
+    },
   ],
   'en': [
     {
